refactoring phase definition to adapt naming style #12 #22

+++ src/microsim/traffic_lights/NEMAController.cpp
@@ -732,7 +732,7 @@ NEMALogic::NEMA_control() {
 
 int NEMALogic::nextPhase(std::vector<int> ring, int currentPhase) {
 
-    int length = ring.size();
+    int length = (int)ring.size();
     int flag = 0;
     int nphase = 0; // next phase
     for (int i = 0; i < length * 2; i++)
